resolution int -> double | rtc_stats_view UI fix

+++ lib/src/model/hms_video_resolution.dart
@@ -1,6 +1,6 @@
 class HMSVideoResolution {
-  final int height;
-  final int width;
+  final double height;
+  final double width;
 
   HMSVideoResolution({required this.height, required this.width});
